+++ src/main/java/maybeweijun/storage/state.txt
@@ -1,5 +1,5 @@
 T | 1 | homework
 D | 0 | submit | 2025-07-23 1921
-E | 0 | party | 1922-10-10 1900 to 1922-11-11 2000
+E | 1 | party | 1922-10-10 1900 to 1922-11-11 2000
 T | 1 | something else
 E | 0 | grad | 2022-08-08 1400 to 2023-02-02 1900
