+++ src/main/java/maybeweijun/task/Deadline.java
@@ -1,0 +1,28 @@
+package maybeweijun.task;
+
+import java.time.LocalDateTime;
+import java.time.format.DateTimeFormatter;
+
+public class Deadline extends Task {
+    protected LocalDateTime by;
+    private static final DateTimeFormatter FORMATTER = DateTimeFormatter.ofPattern("yyyy-MM-dd HHmm");
+    private static final DateTimeFormatter PRINT_FORMATTER = DateTimeFormatter.ofPattern("MMM d yyyy ha");
+    public Deadline(String description, String by) {
+        super(description);
+        this.by = LocalDateTime.parse(by, FORMATTER);
+    }
+
+    public Deadline(String description, LocalDateTime by) {
+        super(description);
+        this.by = by;
+    }
+
+    public LocalDateTime getBy() {
+        return by;
+    }
+
+    @Override
+    public String toString() {
+        return "[D]" + super.toString() + " (by: " + by.format(PRINT_FORMATTER) + ")";
+    }
+}
