+++ src/main/java/maybeweijun/task/Todo.java
@@ -1,0 +1,12 @@
+package maybeweijun.task;
+
+public class Todo extends Task {
+    public Todo(String description) {
+        super(description);
+    }
+
+    public String toString() {
+        return "[T]" + super.toString();
+    }
+
+}
